# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "TechFUL API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "api"
    },
    {
      "url": "http://127.0.0.1:4010",
      "description": "mock"
    }
  ],
  "paths": {
...
    "/v1/selections/{job_id}/{student_id}": {
      "post": {
        "tags": [
          "selections"
        ],
        "summary": "スカウト送信",
        "operationId": "postSelections",
        "parameters": [
          {
            "description": "求人ID",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string",
              "description": "求人ID"
            },
            "name": "job_id",
            "in": "path"
          },
          {
            "description": "学生ID",
            "required": true,
            "schema": {
              "title": "Student Id",
              "type": "string",
              "description": "学生ID"
            },
            "name": "student_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectionsJobIdStudentIdResponse"
                }
              }
            }
          },
          "422": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/messages/{selection_id}": {
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "メッセージ取得（個別）",
        "description": "メッセージタイムラインに表示するメッセージとファイルの一覧を取得する",
        "operationId": "getMessageSelectionId",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Selection Id",
              "type": "string"
            },
            "name": "selection_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageTimelineResponse"
                }
              }
            }
          },
          "422": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "メッセージ、ファイル登録",
        "description": "メッセージorファイルを送信する",
        "operationId": "postMessagesSelectionId",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Selection Id",
              "type": "string"
            },
            "name": "selection_id",
            "in": "path"
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagesResponse"
                }
              }
            }
          },
          "422": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/templates/{template_id}": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "テンプレート取得（特定）",
        "description": "特定のテンプレートを取得する",
        "operationId": "getTemplatesTemplateId",
        "parameters": [
          {
            "description": "テンプレートID",
            "required": true,
            "schema": {
              "title": "Template Id",
              "type": "integer",
              "description": "テンプレートID"
            },
            "name": "template_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplatesTemplateIdResponse"
                }
              }
            }
          },
          "422": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "templates"
        ],
        "summary": "テンプレート更新",
        "description": "テンプレートを更新する",
        "operationId": "putTemplates",
        "parameters": [
          {
            "description": "テンプレートID",
            "required": true,
            "schema": {
              "title": "Template Id",
              "type": "integer",
              "description": "テンプレートID"
            },
            "name": "template_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__routers__v1__templates__put_templates__TemplateData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "templates"
        ],
        "summary": "テンプレート削除",
        "description": "テンプレートを削除する",
        "operationId": "deleteTemplates",
        "parameters": [
          {
            "description": "テンプレートID",
            "required": true,
            "schema": {
              "title": "Template Id",
              "type": "integer",
              "description": "テンプレートID"
            },
            "name": "template_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MessageTimelineResponse": {
        "title": "MessageTimelineResponse",
        "required": [
          "icon",
          "accountName",
          "realName",
          "isDeletedAccount",
          "isScoutRefusalAccount",
          "isScoutAgreed",
          "messages"
        ],
        "type": "object",
        "properties": {
          "icon": {
            "title": "Icon",
            "type": "string",
            "description": "ユーザーアイコン画像URL"
          },
          "accountName": {
            "title": "Accountname",
            "type": "string",
            "description": "ユーザー表示名"
          },
          "realName": {
            "title": "Realname",
            "type": "string",
            "description": "ユーザー本名（スカウト未承諾時は空文字）"
          },
          "isDeletedAccount": {
            "title": "Isdeletedaccount",
            "type": "boolean",
            "description": "退会アカウントか"
          },
          "isScoutRefusalAccount": {
            "title": "Isscoutrefusalaccount",
            "type": "boolean",
            "description": "スカウト拒否アカウントか"
          },
          "isScoutAgreed": {
            "title": "Isscoutagreed",
            "type": "boolean",
            "description": "スカウト承諾しているか"
          },
          "messages": {
            "title": "Messages",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessagesResponse"
            },
            "description": "各メッセージ情報（メッセージがなければ空配列）"
          }
        }
      },
      "MessagesResponse": {
        "title": "MessagesResponse",
        "required": [
          "id",
          "createdAt",
          "messageType",
          "message",
          "isRead",
          "isOwn",
          "sender",
          "icon"
        ],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string",
            "description": "メッセージID"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string",
            "description": "作成日時",
            "format": "date-time"
          },
          "messageType": {
            "title": "Messagetype",
            "type": "integer",
            "description": "1:メッセージ, 2:ファイル"
          },
          "message": {
            "title": "Message",
            "type": "string",
            "description": "メッセージ"
          },
          "isRead": {
            "title": "Isread",
            "type": "boolean",
            "description": "既読か"
          },
          "isOwn": {
            "title": "Isown",
            "type": "boolean",
            "description": "自身のメッセージか"
          },
          "sender": {
            "title": "Sender",
            "type": "string",
            "description": "メッセージ送信者の名前"
          },
          "icon": {
            "title": "Icon",
            "type": "string",
            "description": "メッセージ送信者のアイコン"
          },
          "file": {
            "title": "File",
            "allOf": [
              {
                "$ref": "#/components/schemas/TempfileResponse"
              }
            ],
            "description": "ファイル情報（ファイルの時のみ）",
            "nullable": true
          }
        }
      },
...
      "SelectionsJobIdStudentIdResponse": {
        "title": "SelectionsJobIdStudentIdResponse",
        "required": [
          "createAt",
          "jobId",
          "studentId"
        ],
        "type": "object",
        "properties": {
          "createAt": {
            "title": "Createat",
            "type": "string",
            "description": "スカウト送信キューへの追加日時"
          },
          "jobId": {
            "title": "Jobid",
            "type": "string",
            "description": "選考ID"
          },
          "studentId": {
            "title": "Studentid",
            "type": "string",
            "description": "学生ID"
          }
        }
      },
...
      "SelectionsRequest": {
        "title": "SelectionsRequest",
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "description": "リクエストボディ"
          }
        }
      },
...
      "TempfileResponse": {
        "title": "TempfileResponse",
        "required": [
          "fileName",
          "file",
          "expiresAt",
          "isExpired"
        ],
        "type": "object",
        "properties": {
          "fileName": {
            "title": "Filename",
            "type": "string",
            "description": "ファイル名"
          },
          "file": {
            "title": "File",
            "type": "string",
            "description": "ファイルURL"
          },
          "expiresAt": {
            "title": "Expiresat",
            "type": "string",
            "description": "有効期限",
            "format": "date-time"
          },
          "isExpired": {
            "title": "Isexpired",
            "type": "boolean",
            "description": "有効期限切れか"
          },
          "isDisable": {
            "title": "Isdisable",
            "type": "boolean",
            "description": "削除済みか",
            "default": false
          }
        }
      },
...
      "TemplateUpdateResponse": {
        "title": "TemplateUpdateResponse",
        "required": [
          "templateName",
          "templateSubject",
          "templateBody"
        ],
        "type": "object",
        "properties": {
          "templateName": {
            "title": "Templatename",
            "type": "string",
            "description": "テンプレート名"
          },
          "templateSubject": {
            "title": "Templatesubject",
            "type": "string",
            "description": "件名"
          },
          "jobPositions": {
            "title": "Jobpositions",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/app__schemas__templates__template_update_response__JobPosition"
            },
            "nullable": true
          },
          "templateBody": {
            "title": "Templatebody",
            "type": "string",
            "description": "本文"
          }
        }
      },
      "TemplatesTemplateIdResponse": {
        "title": "TemplatesTemplateIdResponse",
        "required": [
          "useCase",
          "templateName",
          "templateSubject",
          "templateBody"
        ],
        "type": "object",
        "properties": {
          "useCase": {
            "title": "Usecase",
            "minimum": 1.0,
            "type": "integer",
            "description": "用途"
          },
          "templateName": {
            "title": "Templatename",
            "type": "string",
            "description": "テンプレート名"
          },
          "templateSubject": {
            "title": "Templatesubject",
            "type": "string",
            "description": "件名"
          },
          "jobsPosition": {
            "title": "Jobsposition",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/app__schemas__templates__templates_template_id_response__JobPosition"
            },
            "nullable": true
          },
          "templateBody": {
            "title": "Templatebody",
            "type": "string",
            "description": "本文"
          }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "errorCode",
          "message",
          "detail"
        ],
        "type": "object",
        "properties": {
          "errorCode": {
            "title": "Errorcode",
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "title": "Message",
            "type": "string",
            "example": "入力値が不正です"
          },
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorDetail"
            }
          }
        }
      },
      "ValidationErrorDetail": {
        "title": "ValidationErrorDetail",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Loc",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        }
      },
...
      "app__routers__v1__templates__put_templates__TemplateData": {
        "title": "TemplateData",
        "required": [
          "templateName",
          "templateSubject",
          "templateBody"
        ],
        "type": "object",
        "properties": {
          "templateName": {
            "title": "Templatename",
            "type": "string",
            "description": "テンプレート名"
          },
          "templateSubject": {
            "title": "Templatesubject",
            "type": "string",
            "description": "件名"
          },
          "jobPositionIds": {
            "title": "Jobpositionids",
            "type": "array",
            "items": {
              "type": "integer"
            },
            "nullable": true
          },
          "templateBody": {
            "title": "Templatebody",
            "type": "string",
            "description": "本文"
          }
        }
      },
...
      "app__schemas__templates__template_update_response__JobPosition": {
        "title": "JobPosition",
        "type": "object",
        "properties": {
          "jobPositionId": {
            "title": "Jobpositionid",
            "minimum": 0.0,
            "type": "integer",
            "description": "対応する求人ID",
            "nullable": true
          },
          "jobPositionName": {
            "title": "Jobpositionname",
            "type": "string",
            "description": "対応する求人名",
            "nullable": true
          }
        }
      },
...
      "app__schemas__templates__templates_template_id_response__JobPosition": {
        "title": "JobPosition",
        "type": "object",
        "properties": {
          "jobsPositionId": {
            "title": "Jobspositionid",
            "minimum": 0.0,
            "type": "integer",
            "description": "対応する求人ID",
            "nullable": true
          },
          "jobsPositionName": {
            "title": "Jobspositionname",
            "type": "string",
            "description": "対応する求人名",
            "nullable": true
          }
        }
      }
    }
  }
}